Introduce explicit conversions between addresses an numbers.
This is valid on <0.8.0, but will be required on >=0.8.0.

+++ contracts/utils/Create2.sol
@@ -54,6 +54,6 @@ library Create2 {
         bytes32 _data = keccak256(
             abi.encodePacked(bytes1(0xff), deployer, salt, bytecodeHash)
         );
-        return address(uint256(_data));
+        return address(uint160(uint256(_data)));
     }
 }
